test: do not try to kill apm-server if the process did not start

+++ systemtest/apmservertest/server.go
@@ -137,7 +137,10 @@ func NewUnstartedServerTB(tb testing.TB, args ...string) *Server {
 		errc := make(chan error)
 		go func() { errc <- s.Close() }()
 		select {
-		case <-errc:
+		case err := <-errc:
+			if err != nil {
+				tb.Error(err)
+			}
 			close(errc)
 		case <-time.After(10 * time.Second):
 			// Channel receive on errc never happened. Start up a
@@ -431,6 +434,9 @@ func (s *Server) Close() error {
 		return nil
 	}
 	s.closeTracers()
+	if s.cmd.Process == nil {
+		return errors.New("apm server process not started")
+	}
 	if err := interruptProcess(s.cmd.Process); err != nil {
 		s.cmd.Process.Kill()
 	}
